requirements.txt
langgraph>=0.4
langchain>=0.3
langchain-openai>=0.3
langchain-anthropic>=0.3
langchain-community>=0.3
langchain-chroma>=0.2
chromadb>=0.6
mcp[cli]>=1.2
httpx>=0.27
pymupdf>=1.24
pdfplumber>=0.11
ragas>=0.2
streamlit>=1.40
streamlit-agraph>=0.0.45
pyvis>=0.3
python-dotenv>=1.0
arxiv>=2.1
tiktoken>=0.7
langchain-text-splitters>=0.3
